pyproject.toml
[tool.poetry]
name = "gpthub"
version = "0.1.0"
description = "GPTHub — unified AI workspace. OpenAI-compatible backend over any LLM provider: chat, voice, vision, RAG, agentic tools, long-term memory."
authors = ["Alexander Yurlov"]
license = "MIT"
repository = "https://github.com/AlexYrlv/gpthub"
keywords = ["llm", "ai", "rag", "openai-compatible", "multimodal", "agent", "function-calling", "fastapi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115"
uvicorn = "^0.34"
redis = "^5.0"
httpx = "^0.28"
pydantic = "^2.0"
python-multipart = "^0.0.18"
beautifulsoup4 = "^4.12"
pypdf = "^5.0"
python-docx = "^1.1"
duckduckgo-search = "^7.0"
typer = "^0.12"
aiohttp = "^3.11"
aiohttp-socks = "^0.10"
toml = "^0.10"
motor = "^3.3"
pymongo = "^4.8"
config-fastapi = "^1.1"
fastabc = "^4.6"
fastapiabclib = "^4.3"
frpclnt = "^0.5"
commandlinestart = "^0.2.3"
aiomotorengine = "^1.0.1"
red-mng = "^1.1"
aiorediscache = "^4.1"
redis-distlock = "^1.0.1"
aiowebparser = "^1.1.1"
fpiaioact = "^2.0"
redis-rpc-pubsub = "^1.0"
python-pptx = "^1.0.2"
aioboto3 = "^15.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"
ruff = "^0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]
